# Improve performance of `%pat%` (>3x speedup)