# Which issue does this PR close?
Closes #2519

# Rationale for this change
`LIKE %pat%`  is something that occurs a lot in SQL queries. It's also one of the slower queries in the ClickBench benchmarks. 

Results:
```
like_utf8 scalar contains                        
                        time:   [1.9803 ms 1.9828 ms 1.9853 ms]
                        change: [-71.755% -71.321% -70.997%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 3 outliers among 100 measurements (3.00%)
  1 (1.00%) low mild
  2 (2.00%) high mild

Benchmarking nlike_utf8 scalar contains: Warming up for 3.0000 s
Warning: Unable to complete 100 samples in 5.0s. You may wish to increase target time to 9.9s, enable flat sampling, or reduce sample count to 40.
Benchmarking nlike_utf8 scalar contains: Collecting 100 samples in estimated 9.9                                                                                nlike_utf8 scalar contains                        
                        time:   [1.9413 ms 1.9448 ms 1.9488 ms]
                        change: [-71.544% -71.455% -71.344%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 10 outliers among 100 measurements (10.00%)
  1 (1.00%) low mild
  6 (6.00%) high mild
  3 (3.00%) high severe
```

# What changes are included in this PR?

<!--
There is no need to duplicate the description in the issue here but it is sometimes worth providing a summary of the individual changes in this PR.
-->


# Are there any user-facing changes?


<!--
If there are user-facing changes then we may require documentation to be updated before approving the PR.
-->

<!---
If there are any breaking changes to public APIs, please add the `breaking change` label.
-->